[US-006] Implement MarkdownParser service
- Add MarkdownParser struct for YAML frontmatter parsing
- Extract shortcut (required), category (optional), enabled (default true)
- Add parse() for text and parse(fileAt:) for file URLs
- Add generateMarkdown() for creating snippet files
- Refactor Snippet to use MarkdownParser (separation of concerns)
